basic boolean conversion and interpretation behavior
The idea here is to interpret incoming values as booleans in a friendly,
flexible way and be able to easily compare them to other booleans (both
`MuchBoolean` and `TrueClass` and `FalseClass`).  Use `convert` to just
convert a given value to an actual boolean value using much-boolean's
"friendly" interpretation logic.  Create instances using `new` to track
a given values interpreted boolean value and use as a boolean in
comparisons, etc.

Note: this was extracted from the ns-options gem with a few differences:

* no `actual=` writer (this was only needed b/c of ns-options behavior)
* no `returned_value` method (this was only needed b/c of ns-options
  behavior)
* support for "yes/no" type values
* support for additional "true/false" type values
* exposed the underlying conversion logic in a `convert` singleton so
  you don't *have* to build an instance and get its `actual` value
  if all you care about is the actual boolean value