add in method to convert boolean values back to human-friendly values
This is the opposite of the `convert` method, provided in a variety
of formats based on the `FALSE_VALUES` that `convert` supports.  The
tests try to illustrate that converting a human-friendly value generated
with this API results in the boolean used to generate the human-friendly
value.

The goal here is to provide an complimentary API to `convert` for
generating human friendly values `convert` can process.