hotfix: interpret empyt-string values like nil values
This is somewhat of a judgement call but I think it is a less
surprising way to treat empty string than seeing it is true.  We
are already off of the default ruby way of interpreting objects as
booleans (that is the point of the gem).  Plus the string representation
of nil is empty string so this really is consistent with the string
representations of objects map as those objects themselves would.

/cc @jcredding